take addObject from actions and place in helpers, remove containers
