+++ src/components/Parent/AddSitter.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import { addObject } from '../../helpers/helper'
 
 export default class AddSitter extends Component {
   constructor (){
@@ -45,7 +46,7 @@ export default class AddSitter extends Component {
           />
           <button type='submit'
               className="form submit add-btn"
-              onClick={()=>this.props.addSitter(this.state)}>
+              onClick={()=> addObject(this.state, 'sitters')}>
               Add Sitter
             </button>
         </form>
