+++ src/containers/AddChildContainer.js
@@ -1,6 +1,4 @@
 import {connect} from 'react-redux';
-import { addObject } from '../actions'
-
 import AddChild from '../components/Children/AddChild'
 
 const mapStateToProps = (state) => {
@@ -9,10 +7,6 @@ const mapStateToProps = (state) => {
   }
 }
 
-const mapDispatchToProps = (dispatch) => {
-  return {
-    addChild: (child) => dispatch(addObject(child, 'children'))
-  }
-}
 
-export default connect(mapStateToProps, mapDispatchToProps)(AddChild)
+
+export default connect(mapStateToProps, null)(AddChild)
